event.py를 여러 파일로 분리
