+++ engine/focus.py
@@ -3,7 +3,8 @@ from enum import Enum
 
 import pygame
 
-from engine.event import Event, EventEmitter
+from engine.events.emitter import EventEmitter
+from engine.events.event import Event
 
 
 class FocusMoveDirection(Enum):
