+++ engine/gameobject.py
@@ -2,7 +2,8 @@ import abc
 
 import pygame
 
-from engine.event import Event, EventEmitter
+from engine.events.emitter import EventEmitter
+from engine.events.event import Event
 
 
 class GameObject(EventEmitter, abc.ABC):
