+++ game/lobby/lobbyscene.py
@@ -1,7 +1,7 @@
 import pygame
 
 from engine.button import Button
-from engine.event import Event
+from engine.events.event import Event
 from engine.focus import FocusMoveDirection
 from engine.layout import LayoutAnchor
 from engine.scene import Scene
